+++ src/main/java/racingcar/controller/Movement.java
@@ -1,8 +1,8 @@
 package racingcar.controller;
 
 import java.util.List;
-import racingcar.model.Car;
-import racingcar.model.Position;
+import racingcar.domain.Car;
+import racingcar.domain.Position;
 import racingcar.util.RandomGenerator;
 
 public class Movement {
