+++ src/test/java/racingcar/MovementTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import racingcar.controller.Movement;
-import racingcar.model.Position;
+import racingcar.domain.Position;
 
 
 public class MovementTest {
