+++ src/test/java/racingcar/WinnerTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import racingcar.controller.Winner;
-import racingcar.model.Car;
+import racingcar.domain.Car;
 
 public class WinnerTest {
